move kerning_for_non_ligated_sequences out of googlefonts directory
